# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Vexel Health API (LIMS-first)
  version: 0.1.0
  description: |
    Contract-first OpenAPI spec for Vexel Health Platform.
    All admin endpoints require bearerAuth.
    All requests should include X-Correlation-ID header.
    Tenant is resolved server-side (Host header in prod, x-tenant-id in dev when enabled).
servers:
  - url: /api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
        correlationId:
          type: string
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    Patient:
      type: object
      required: [id, tenantId, firstName, lastName, mrn, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
          nullable: true
        ageYears:
          type: integer
          nullable: true
        gender:
          type: string
          nullable: true
        mrn:
          type: string
        mobile:
          type: string
          nullable: true
        cnic:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Encounter:
      type: object
      required: [id, tenantId, patientId, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        patientId:
          type: string
        encounterCode:
          type: string
          nullable: true
          description: "LIMS Order ID e.g. VXL-2602-001"
        status:
          type: string
          enum: [registered, lab_ordered, specimen_collected, specimen_received, partial_resulted, resulted, verified, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        patient:
          $ref: '#/components/schemas/Patient'
        labOrders:
          type: array
          items:
            $ref: '#/components/schemas/LabOrder'
        specimenItems:
          type: array
          items:
            $ref: '#/components/schemas/SpecimenItem'
    LabOrder:
      type: object
      required: [id, tenantId, encounterId, testId, status, resultStatus, priority, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        encounterId:
          type: string
        testId:
          type: string
        testNameSnapshot:
          type: string
          nullable: true
        test:
          type: object
          description: Resolved test definition (populated when available)
          nullable: true
          properties:
            id:
              type: string
            code:
              type: string
            name:
              type: string
            specimenType:
              type: string
              nullable: true
            loincCode:
              type: string
              nullable: true
        status:
          type: string
          enum: [ordered, specimen_collected, processing, resulted, verified, cancelled]
        resultStatus:
          type: string
          enum: [PENDING, SUBMITTED]
          default: PENDING
        priority:
          type: string
          enum: [routine, stat, urgent]
        submittedAt:
          type: string
          format: date-time
          nullable: true
        submittedById:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        specimen:
          $ref: '#/components/schemas/Specimen'
        results:
          type: array
          items:
            $ref: '#/components/schemas/LabResult'
    Specimen:
      type: object
      required: [id, tenantId, labOrderId, barcode, type, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        labOrderId:
          type: string
        barcode:
          type: string
        type:
          type: string
        collectedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, collected, received, rejected]
        createdAt:
          type: string
          format: date-time
    LabResult:
      type: object
      required: [id, tenantId, labOrderId, value, enteredAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        labOrderId:
          type: string
        parameterId:
          type: string
          nullable: true
        parameterNameSnapshot:
          type: string
          nullable: true
        value:
          type: string
        unit:
          type: string
          nullable: true
        referenceRange:
          type: string
          nullable: true
        flag:
          type: string
          enum: [normal, high, low, critical]
          nullable: true
        locked:
          type: boolean
          default: false
        enteredAt:
          type: string
          format: date-time
        enteredById:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        verifiedBy:
          type: string
          nullable: true
    SpecimenItem:
      type: object
      required: [id, tenantId, encounterId, catalogSpecimenType, status, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        encounterId:
          type: string
        catalogSpecimenType:
          type: string
          description: "Exact specimen type e.g. 'EDTA Blood', 'Urine'"
        status:
          type: string
          enum: [PENDING, COLLECTED, POSTPONED, RECEIVED]
        barcode:
          type: string
          nullable: true
        collectedAt:
          type: string
          format: date-time
          nullable: true
        collectedById:
          type: string
          nullable: true
        postponedAt:
          type: string
          format: date-time
          nullable: true
        postponeReason:
          type: string
          nullable: true
        receivedAt:
          type: string
          format: date-time
          nullable: true
        receivedById:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TestParameter:
      type: object
      description: A catalog parameter definition with current value for results entry
      properties:
        parameterId:
          type: string
        name:
          type: string
        unit:
          type: string
          nullable: true
        dataType:
          type: string
          enum: [number, text, select, boolean]
          default: number
        allowedValues:
          type: array
          items:
            type: string
          nullable: true
          description: For dataType=select
        referenceRange:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
          description: Current entered value (if any)
        flag:
          type: string
          enum: [normal, high, low, critical]
          nullable: true
        locked:
          type: boolean
          default: false
          description: True if test submitted and value is non-empty
    OrderedTestDetail:
      type: object
      description: Full detail for one test (LabOrder) including parameter schema and current values
      required: [id, encounterId, testId, testName, resultStatus]
      properties:
        id:
          type: string
          description: LabOrder ID (orderedTestId)
        encounterId:
          type: string
        encounterCode:
          type: string
          nullable: true
        testId:
          type: string
        testName:
          type: string
        resultStatus:
          type: string
          enum: [PENDING, SUBMITTED]
        submittedAt:
          type: string
          format: date-time
          nullable: true
        patient:
          $ref: '#/components/schemas/Patient'
        specimenStatus:
          type: string
          description: "Status of the specimen for entry gating"
          nullable: true
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/TestParameter'
    OrderedTestSummary:
      type: object
      description: Row item for results worklist
      required: [id, encounterId, testName, resultStatus, createdAt]
      properties:
        id:
          type: string
          description: LabOrder ID
        encounterId:
          type: string
        encounterCode:
          type: string
          nullable: true
        testId:
          type: string
        testName:
          type: string
        resultStatus:
          type: string
          enum: [PENDING, SUBMITTED]
        submittedAt:
          type: string
          format: date-time
          nullable: true
        filledCount:
          type: integer
          description: Number of parameters with values entered
        totalCount:
          type: integer
          description: Total number of parameters in this test
        specimenStatus:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        patient:
          $ref: '#/components/schemas/Patient'
...
    SampleCollectionEncounter:
      type: object
      description: Row item for sample collection worklist
      required: [encounterId, pendingCount, totalCount, createdAt]
      properties:
        encounterId:
          type: string
        encounterCode:
          type: string
          nullable: true
        pendingCount:
          type: integer
        totalCount:
          type: integer
        testsCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        patient:
          $ref: '#/components/schemas/Patient'
        specimenItems:
          type: array
          items:
            $ref: '#/components/schemas/SpecimenItem'
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      summary: List patients (tenant-scoped)
      tags: [Patients]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: lastName
          in: query
          schema:
            type: string
        - name: mrn
          in: query
          schema:
            type: string
        - name: mobile
          in: query
          schema:
            type: string
          description: Search by mobile number (exact or partial match)
      responses:
        "200":
          description: Patients list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      operationId: createPatient
      summary: Create a patient record
      tags: [Patients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName]
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                dateOfBirth:
                  type: string
                  format: date
                ageYears:
                  type: integer
                gender:
                  type: string
                mobile:
                  type: string
                cnic:
                  type: string
                address:
                  type: string
                mrn:
                  type: string
                  description: "If omitted, server generates from tenant prefix"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        "409":
          description: MRN already exists in tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /patients/{patientId}:
    get:
      operationId: getPatient
      summary: Get patient by ID
      tags: [Patients]
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updatePatient
      summary: Update patient record
      tags: [Patients]
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                dateOfBirth:
                  type: string
                  format: date
                gender:
                  type: string
      responses:
        "200":
          description: Updated patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Encounters ───────────────────────────────────────────────────────────
...
  /sample-collection/worklist:
    get:
      operationId: getSampleCollectionWorklist
      summary: List encounters with pending specimen items
      tags: [SampleCollection]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, POSTPONED, RECEIVED]
            default: PENDING
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: "Default: 3 days ago"
        - name: toDate
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
          description: Search by MRN, patient name, or encounter code
      responses:
        "200":
          description: Sample collection worklist
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SampleCollectionEncounter'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /encounters/{encounterId}:collect-specimens:
    post:
      operationId: collectSpecimens
      summary: Batch collect specimen items for an encounter
      tags: [SampleCollection]
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                specimenItemIds:
                  type: array
                  items:
                    type: string
                  description: "IDs of SpecimenItems to collect; empty = collect all PENDING"
      responses:
        "200":
          description: Specimens collected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        "409":
          description: Invalid state or no pending specimens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /encounters/{encounterId}:postpone-specimen:
    post:
      operationId: postponeSpecimen
      summary: Postpone a specimen item with a reason
      tags: [SampleCollection]
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [specimenItemId, reason]
              properties:
                specimenItemId:
                  type: string
                reason:
                  type: string
                  minLength: 3
      responses:
        "200":
          description: Specimen postponed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecimenItem'
        "409":
          description: Specimen not in postponable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /encounters/{encounterId}:receive-specimens:
    post:
      operationId: receiveSpecimens
      summary: Batch receive specimen items (requires lims.operator.sample.receiveSeparate.enabled)
      tags: [SampleCollection]
      parameters:
        - name: encounterId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                specimenItemIds:
                  type: array
                  items:
                    type: string
                  description: "IDs of SpecimenItems to receive; empty = receive all COLLECTED"
      responses:
        "200":
          description: Specimens received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        "403":
          description: Feature flag disabled for tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

  # ─── Results (test-level) ─────────────────────────────────────────
  /results/tests/pending:
    get:
      operationId: getPendingTests
      summary: List tests pending result entry
      tags: [Results]
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by MRN, patient name, order code, or test name
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Pending tests worklist
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderedTestSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /results/tests/submitted:
    get:
      operationId: getSubmittedTests
      summary: List tests submitted for verification
      tags: [Results]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Submitted tests list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderedTestSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /results/tests/{orderedTestId}:
    get:
      operationId: getOrderedTestDetail
      summary: Get test detail with parameter schema and current values
      tags: [Results]
      parameters:
        - name: orderedTestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ordered test with parameter values and lock state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderedTestDetail'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...